+++ app/helpers/project_items_helper.rb
@@ -25,6 +25,6 @@ module ProjectItemsHelper
 
   def project_item_ready?(project_item)
     project_item_change = project_item.project_item_changes.where(status: :ready).pluck(:created_at)
-    project_item_change.empty? ? false : (project_item_change.first < RATE_PERIOD)
+    project_item_change.empty? ? false : (project_item_change.first < RATE_AFTER_PERIOD.ago)
   end
 end