+++ config/initializers/constants.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-# time after which mail is send
-RATE_PERIOD = 3.months.ago
 # time after which user can rate service that starts when service become ready
-RATE_AFTER_PERIOD = 3.months
+# default value is set to 90 days, ENV variable should represent number of days
+RATE_AFTER_PERIOD = ENV["RATE_AFTER_PERIOD"].present? ? (ENV["RATE_AFTER_PERIOD"]).to_i.days : 90.days