+++ CHANGELOG.md
@@ -130,6 +130,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Add new supported JIRA states (rejected, waiting for response) (@michal-szostak)
 - ASSET_HOST and ASSET_PROTOCOL for mailer config (@kmarszalek) 
 - Additional field mapping to JIRA order (@michal-szostak)
+- Possibility to setup RATE_AFTER_PERIOD ENV variable (@kmarszalek)
 
 ### Changed
 - Upgrade Sprockets gem to avoid CVE-2018-3760 vulnerability (@mkasztelnik)